move map related files in package dir "map"
