+++ meson.build
@@ -11,10 +11,10 @@ src_files = files(
   'include/termbox2.c',
   'src/main.c',
   'src/game.c',
-  'src/map.c',
-  'src/map_mode.c',
-  'src/map_generator.c',
-  'src/map_populator.c',
+  'src/map/map.c',
+  'src/map/map_mode.c',
+  'src/map/map_generator.c',
+  'src/map/map_populator.c',
 )
 
 debug_files = files(
