+++ src/map/map_generator.h
