+++ src/map/map_mode.c
