+++ src/map/map_mode.h
